implement projects section and fix dev servers section
